graphiti_memory/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphiti-memory"
version = "0.1.0"
description = "Personal memory layer using Graphiti temporal knowledge graph over Claude conversation history."
requires-python = ">=3.11"
dependencies = [
    "graphiti-core",
    "anthropic",
    "neo4j",
    "click",
    "pyyaml",
    "rich",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
memory-ingest = "src.ingest:main"
memory-chat = "src.chat:main"
memory-inspect = "src.inspect:main"
